CRAYSAT-60: Refactor subcommands into subpackages
Refactor the existing code for the `showrev` subcommand into its own
subpackage to provide a model for other subcommands to follow that will
keep code for each subcommand organized into its own subpackage. This
should make it easier to collaborate and prevent the code in our main
method from growing too long and unwieldy.

Use a dictionary that maps from subcommand name to the function
implementing the subcommand. Allow parse_args to catch any invalid
subcommand names and print an error. Print help when not given any
subcommands and exit non-zero.

Also implement a stub for unit tests for the sat.showrev.system module.
We will populate this stub with actual unit tests in a separate JIRA.
Remove the `test_cli.py` unit test file stub as it is outdated with the
new project structure.

Also update the `setup.py` script to reference the location of the new
main method, which has moved into a module named main at the top level
of the sat package.

Test Description:
Installed in a virtual environment on spotted cow and ran `showrev` with
different combinations of options. Also verified that the program
returned an error and printed usage information when run without a
subcommand or when run with an invalid subcommand.

Ran unit tests and verified that they passed.