CRAYSAT-31: Showrev system
DST implemented a release file where more value streams are stored. This
reports the versions stored in that file, as well as other system
information that xtshowrev used to report - resulting in a more accurate
replication of that behavior...

Though this command is not complete yet - and likely won't be by 1.0.0
as Shasta has no concept of how to report this information.
Jiras have been filed for the following features.

  - Site name
  - Serial Number
  - System type
  - Interconnect
  - Install date
  - System name

This change also includes unit tests for get_sles_version.

Test Description:

I ran the --system operation from sat showrev and saw my expected
output when I redirected the utility to a custom 'value-streams' file. I
also ran all unit tests on my developer machine and spottedcow and they
passed.