CRAYSAT-144: Port configuration to TOML format.
This change translates the configuration format into TOML. TOML is
used by the Cray CLI on Shasta systems, and is easier to implement
than the built-in Python ConfigParser library.

This change also changes the way in which get_config_value() is
used. Since our config file follows a basic two-tiered hierarchy
to give us a little bit of organization, get_config_value() now
accepts a "query string" in the form '<section>.<option>' instead
of an optional 'section' parameter. Thus clients of this method
are expected to always specify the section name when querying
options. In the future, we may want to have multiple-layered
hierarchies, but that is a feature for another day.

This change also makes it so that sat.toml is generated automatically
when RPMs are built for cray-sat. Thus, this makes it such that there
is a 'single source of truth' regarding the default configuration;
there can be no discrepancy between the spec in sat/config.py and
etc/sat.toml.

Test Description:
1. Unit tests have been updated; use nosetests to check for any
   regressions.
2. Test command line functionality relating to features that use
   the configuration file.
3. Run `./generate_default_config.py sat/config.py`
   and verify that the file is valid TOML which reflects the
   OptionSpec provided in sat/config.py.