CRAYSAT-142: Implement output filtering.
This commit adds utilities for filtering output against specific
queries specified on the command line. This is done using a simple
'query language' which allows for comparisons of attributes against
values, and combinations of these using logical boolean
operators ('and' and 'or'.) Values can be wildcards, string
literals (enclosed in single- or double-quotes), or numbers. A simpler
version of the query language was also implemented which is roughly
equivalent to the filtering mechanism used in XC products such as
`cnode` or `cmap`.

This commit also adds tests for the new functionality, as well as an
ArgumentParser which can be used as a parent parser for subcommand
parsers so that the filtering options can be generalized across these
subcommands.

Test Description:
1. Run unit tests to ensure they pass.
2. At the moment, this code has not yet been integrated with the
   tools as they currently function; however, once they are, filtering
   can be tested on 'real hardware'.