CRAYSAT-149: Add unit tests for sat.system and test stubs for sat.cli.hwinv
Add unit tests for all modules in the `sat.system` package except for
the `system` module. Add a stub test file for the `system` module to
force the coverage of that module to be reported when unit tests are
run. The stub should be filled in with unit tests later.

Fix a minor bug where `BaseComponent.get_unique_child_vals` would fail
trying to join non-str values if the unique values were not of type str.

Fix minor docstring error in component.py.  Remove a couple unused
imports as well.

Implement a `get_component_raw_data` function in a `component_data`
module in `tests.system` so that it can be used by multiple tests that
need to create raw data for creating component instances.

Add stub unit test files for the `sat.cli.hwinv` subpackage. There are
no real tests here yet, but it does force the coverage module to
accurately report on coverage. References to JIRAs to implement
remaining tests are included in code comments.

Test Description:
Unit tests passing with 100% coverage on all of the system package
except the system module within it.