CRAYSAT-375: Fus to fas
SAT Firmware command now supports both the FUS and the FAS.

This was accomplished by creating two classes to represent clients of
each respective service. The classes have identical interfaces. Call
the 'create_firmware_client' function, which will return the correct
class based on which service is available.

Test description:

I tested `sat firmware` with no opts, --snapshots, and --xnames on my
locally running FAS, as wel as on monster, a system on which I created a
snapshot using a custom sat.toml, and the python3 interpreter to
directly import the create_firmware_client. I used this client to issue
a get request to 'snapshot/botest', which creates a snapshot on the FUS.