CRAYSAT-505: Only add log handlers for sat package
Only add log handlers for the logger named 'sat', i.e. the top-level
logger for the `sat` package in the logger hierarchy. This will prevent
us from logging excessive information from the other 3rd-party packages
that we import. In particular, this will prevent us from emitting the
token in plain text to stderr and to the log file when `sat` is executed
with `--loglevel debug`. That message was coming from requests_oauthlib.

This will also prevent other sensitive information from leaking to log
files if there are other 3rd-party packages we use that log sensitive
information.

Note that if any 3rd-party package we are using uses the `logging`
package and gets a logger for their package but does not add a
`NullHandler` to their top-level logger, then it will still be possible
for those log messages to be emitted by the "handler of last resort"
which is used when no handlers are configured. By default this will log
messages to stderr if they are a level of WARNING or above. We might
want to reconsider this behavior if it becomes a problem, but so far I
have not seen this be a problem.

Add assertions to unit tests to assert that we are configuring the
logger named `sat` and that we set the level on the logger to DEBUG.

Also, fix a missing space in a log message.

Update the changelog with a version in preparation of merge to master.

Test Description:
Unit tests pass. pycodestyle reports no violations.

Tested on gamora-ncn-w001 and verified only messages from `sat` package
are logged to console and to file with log level set to debug.

Verified that RPM built and had new version.