CRAYSAT-482: Implement BOS shutdowns in `sat bootsys shutdown`
Implement automatic BOS shutdowns of all the compute nodes and User
Access Nodes (UANs) in `sat boosys shutdown`. By default, this will use
the session template that has a name of the form `cle-x.y.z` to shut
down the CLE compute nodes and a session template named `uan` to shut
down the UANs. These are the default session templates for these nodes.
Both these values are configurable in the config file or on the command
line, with the command line taking precedence.

Once the appropriate session templates are identified, the code will
attempt to identify whether the nodes in the session templates are all
already in the desired 'Off' state before attempting to create a session
with the 'shutdown' operation against the session templates. If there
are any nodes in a session template that are not in the desired state,
it will create a session against that session template with the
'shutdown' operation. In the future, it would be possible to be more
granular and only shutdown the nodes that need to be shutdown with the
`limit` parameter when creating the BOS session. If errors occur at this
stage, by default the operation will be aborted, but it is possible for
the user to change this behavior with a config file option and
command-line option.

Once the session templates that need to have a shutdown operation
performed against them are identified, the code creates a session with
the 'shutdown' operation against each of them, and it does so in a
separate thread for each session template. This allows us to keep the
main thread of execution free for quickly checking on the status of the
sessions. In the future, we may want to allow the user to request more
info about the running sessions (e.g. by typing 's').

The sessions are watched until they complete or until they time out, and
then the command exits with an appropriate exit code indicating success
(0) or failure (1).

Note that this code currently has a prompt that will make the user enter
'yes' in order to proceed with BOS shutdowns to prevent people from
accidentally shutting down all computes. In the future we will probably
want a non-interactive mode as well.

Note that there are two implementations of monitoring BOS session
status. The one that uses the BOS status endpoint does not work due to a
BOS bug right now, so use a method that uses  `kubectl wait`, `kubectl
get`, and `kubectl logs` commands to determine if the BOS session has
completed and to see if it was successful for the time being.

Add a `json` parameter to APIGatewayClient.post method because it
automatically includes the "Content-Type": "application/json", which is
required by BOS when creating a session. I'm not sure why it is not
required by the other services to which we are making POST requests
already.

Fix the RPM build by making sure that the `generate_default_config.py`
script can find the `sat` package, because an import from the `sat`
package has been added to the `sat.config` module.

Test Description:
Tested on mug-ncn-w001 in the following scenarios:
* cle_bos_template set to a session template containing two nodes and
  uan_bos_template set to empty string to disable it.
	* regular BOS timeout to allow sessions to complete
	* shortened BOS timeout to force a timeout
* cle_bos_template set to a session_template containing one node and
  uan_bos_templte set to a session_template containing one node.
	* regular BOS timeout to allow sessions to complete
	* shortened BOS timeout to force a timeout

All existing unit tests passing and pycodestyle reports no violations.
Some unit tests have been added, but the `sat.cli.bootsys.bos` module
still needs significantly more coverage. A follow-up JIRA will be filed
to add to these.