CRAYSAT-569: Log begin time, end time and duration of bootsys stages
Add a context manager that logs a message when it enters the context
manager, when it exits the context manager, and includes the duration of
the context, so that we can accurately report on how long each stage of
`sat bootsys` takes.

Added unit tests for new `BeginEndLogger` context manager.

Test Description:
Tested on noname and verified that logging occurred as expected at the
beginning and end of contexts where the context manager is used.

Added unit tests pass. pycodestyle reports no violations.