CRAYSAT-250: Use new API for sat diag
This commit changes the sat diag command to use the
new HMS hardware access API, named Fox.

Add a FoxClient class for interfacing with the HTTP API

Modify the RunningDiagPool and DiagStatus classes to work
with the new flow of using the Fox API, which now deals in
jobs that can be launched across multiple controllers instead
of sending requests to individual controllers. This requires
some changes to the logic in order to handle error responses
to POST requests to the controller (HMJTD).

Add unit tests for RunningDiagPool and FoxClient.

Test Description:
Ran the following commands:
  sat diag -x x3000c0r24b0 runMemTester -h
  sat diag -x x3000c0r24b0 runMemTester --option-that-doesnt-exist
  sat diag -x x3000c0r24b0 diagThatDoesntExist
I also ran a couple commands against multiple xnames;
  sat diag -x x3000c0r24b0 -x x3000c0s1b0 runMemTester -h
  sat diag -x x3000c0r24b0 -x x3000c0s1b0 runMemTester --option-that-doesnt-exist
  sat diag -x x3000c0r24b0 -x x3000c0s1b0 diagThatDoesntExist
Finally I tested that error handling works given a diag that works for
one type of controller but not for another:
  sat diag -x x3000c0r24b0 -x x3000c0s1b0 runDiagPfgConfig -h