CRAYSAT-1172: Add info about failed API requests to APIError
Add more information about failed API requests to the APIError that is
raised in APIGatewayClient._make_req when a request gets a response from
the API with an HTTP error status code.

Obtain this information from the fields that are present in the
"application/problem+json" content type as defined in RFC 7807. Many of
the Shasta service APIs provide information about failed requests using
this error response format.

Test Description:
Tested in `sat bootprep` when requests to PUT configurations to the CFS
API failed.

Unit tests pass.