CRAYSAT-1082, CRAYSAT-1085: Create CFS configurations in `sat bootprep`
Add functionality to `sat bootprep` to create CFS configurations that
are defined in the `bootprep` input file. This includes support for
configuration layers defined directly with a git URL and commit hash or
branch as well as configuration layers defined with the name of a
product and a version and/or a branch.

Add options to `sat bootprep` to control the behavior when
configurations which are to be created already exist.

Add a `--dry-run` option to do just a dry run of `sat bootprep` without
actually creating any of the configurations.

Add `--save-files` and `--output-dir` options to allow the admin to
output the files that can be used with direct API requests to CFS if
desired.

Make a change to pass JSON-encoded data in a PUT request in the
`APIGatewayClient` class, which is required for making PUT requests to
CFS to create or update CFS configurations.

Test Description:
All unit tests pass. pycodestyle passes on newly added or modified code.

Tested on thanos. Tested behavior of prompting and
`--skip-existing-configs` and `--overwrite-configs` options when a
config exists. Tested behavior when CFS fails to create the
configuration due to issues in CFS.