CRAYSAT-1198: Implement `rename_image` method on `IMSClient`
Implement a `rename_image` method on the `IMSClient` that renames an
image by performing a deep copy of the image and then a delete of the
old image.

A deep copy of the image means that the each artifact specified in the
image manifest file is copied to a new location in S3, a new manifest is
created that references each of these new artifacts, that new manifest
is uploaded to S3, and then the IMS image is updated to refer to that
new manifest.

Updated `APIGatwayClient.patch` to allow specifying json.  The `patch`
method of the `APIGatewayClient` class no longer requires the `payload`
keyword argument, so adjust the unit tests to no longer special case the
subtests for that method.

Test Description:
Unit tests pass. Pycodestyle passes.

Tested image copy and rename using a test python script that calls the
new methods directly with an existing image ID and a new name.

After a copy or rename of the image using the script, the manifest for
the image was downloaded from S3, the artifacts specified in the
manifest were downloaded and described with a GET request, and the
following conditions were validated:

* The e_tag of each artifact (kernel, rootfs, initrd) in S3 matched the
  e_tag specified in the manifest file.
* The md5sum specified in the metadata for the manifest file matched the
  md5sum of the downloaded manifest file.
* The md5sum specified in the metadata for each artifact matched the
  md5sum of the downloaded file.
* The md5sum of each artifact in the new image matched the md5sum of the
  artifacts from the old image prior to copy/rename.
* The types of the artifacts reported by the `file` command were
  correct.
* Metadata of each artifact correctly referred back to the name and ID
  of the associated image in IMS.