CRAYSAT-1174: Look up commit hash by branch name when needed
This change adds functionality to look up commit hashes which correspond
to the given remote branches in a CFS configuration layer in a bootprep
input file when CSM doesn't support branch names. (CSM support for
branch names is indicated by a simple boolean flag, which can be toggled
once CSM 1.2 is available to SAT. This will return SAT to the old
behavior. In the meantime, the changes in this commit are toggled off
for unit tests which test the intended behavior.)

Test Description:
Run unit tests, and test on thanos using specially constructed bootprep
input files which reference a test repository with no-op configuration
playbooks.