CRAYSAT-1265: Automatically adjust container status width
This change adds a simple calculation to find the maximum length of the
container names for the configuration containers when configuring images
created by `sat bootprep`. The lengths of the width of the container
names in the status log lines no longer need to be hard-coded.

Test Description:
Run manually on an internal system, and ensure that configuration logs
are formatted with the proper widths (i.e not padded more than the
longest container name.)