CRAYSAT-564: Add xname limiting to sat bootsys commands
This change adds a new command line argument, `--bos-limit`, which
passes through the given filter parameter to the BOS `limit` parameter
in the POST body when creating a BOS session to boot or shutdown the
given nodes.

This change also adds a command line argument, `--recursive`, to `sat
bootsys` which causes xnames passed to `--bos-limit` to be expanded into
their constituent node component xnames. For instance, a slot xname
passed to `--bos-limit` would be replaced with the xnames of all node
components in that slot. The user is prompted to confirm the number of
affected xnames when using this option.

Test Description:
Run unit tests. Run `sat bootsys shutdown` and `sat bootsys boot`
operations on an internal system, testing with individual blades.